src/test/java/seedu/giatros/storage/StorageManagerTest.java
package seedu.giatros.storage;

import static org.junit.Assert.assertEquals;
import static org.junit.Assert.assertNotNull;
import static seedu.giatros.testutil.TypicalPatients.getTypicalGiatrosBook;

import java.nio.file.Path;

import org.junit.Before;
import org.junit.Rule;
import org.junit.Test;
import org.junit.rules.TemporaryFolder;

import seedu.giatros.commons.core.GuiSettings;
import seedu.giatros.model.GiatrosBook;
import seedu.giatros.model.ReadOnlyGiatrosBook;
import seedu.giatros.model.UserPrefs;

public class StorageManagerTest {

    @Rule
    public TemporaryFolder testFolder = new TemporaryFolder();

    private StorageManager storageManager;

    @Before
    public void setUp() {
        JsonGiatrosBookStorage addressBookStorage = new JsonGiatrosBookStorage(getTempFilePath("ab"));
        JsonUserPrefsStorage userPrefsStorage = new JsonUserPrefsStorage(getTempFilePath("prefs"));
        storageManager = new StorageManager(addressBookStorage, userPrefsStorage);
    }

    private Path getTempFilePath(String fileName) {
        return testFolder.getRoot().toPath().resolve(fileName);
    }


    @Test
    public void prefsReadSave() throws Exception {
        /*
         * Note: This is an integration test that verifies the StorageManager is properly wired to the
         * {@link JsonUserPrefsStorage} class.
         * More extensive testing of UserPref saving/reading is done in {@link JsonUserPrefsStorageTest} class.
         */
        UserPrefs original = new UserPrefs();
        original.setGuiSettings(new GuiSettings(300, 600, 4, 6));
        storageManager.saveUserPrefs(original);
        UserPrefs retrieved = storageManager.readUserPrefs().get();
        assertEquals(original, retrieved);
    }

    @Test
    public void addressBookReadSave() throws Exception {
        /*
         * Note: This is an integration test that verifies the StorageManager is properly wired to the
         * {@link JsonGiatrosBookStorage} class.
         * More extensive testing of UserPref saving/reading is done in {@link JsonGiatrosBookStorageTest} class.
         */
        GiatrosBook original = getTypicalGiatrosBook();
        storageManager.saveGiatrosBook(original);
        ReadOnlyGiatrosBook retrieved = storageManager.readGiatrosBook().get();
        assertEquals(original, new GiatrosBook(retrieved));
    }

    @Test
    public void getGiatrosBookFilePath() {
        assertNotNull(storageManager.getGiatrosBookFilePath());
    }

}
